tests: Amend buggy workaround for distcheck to not break check.

+++ tests/API/XCCDF/unittests/test_report_check_with_empty_selector.sh
@@ -10,13 +10,17 @@ stderr=$(mktemp -t ${name}.out.XXXXXX)
 
 	# Workaround trac#245 for distcheck
 	oval=${name}.oval.xml.result.xml
-	ln -s $srcdir/$oval $oval
+	if [ ! -f "$oval" ]; then
+		ln -s $srcdir/$oval $oval
+	fi
 	# Workaround end
 
 $OSCAP xccdf generate --format html report --output $result $srcdir/${name}.xccdf.xml.result.xml 2> $stderr
 
 	# Workaround
-	rm -rf $oval
+	if [ -L "$oval" ]; then
+		rm -rf $oval
+	fi
 	unset oval
 	# Workaround end
